reference count bugfix
Previously the ref counts on the numeric indifferent prefs were being
decremented whenever the ref count of the associated acceptable preference was
decremented. However, this leads to the numeric indifferent prefs being
deallocated before the acceptable pref is deallocated, since the acceptable
pref is usually referenced by many pointers and I don't increment the ref
counts on the associated numeric indifferent prefs whenever the ref count on
the acceptable pref is incremented. This fix changes the code so that the
numeric indifferent preferences are only deallocated when the acceptable pref
they're linked to is deallocated.