Added ability to use cross validation to choose number of components in PCR
This was in response to some models choosing to use all components and
getting betas with very large norms that resulted in ridiculous
out-of-training predictions. Unfortunately the cross validation doesn't
seem to work too well; it chooses too few components resulting in
inaccurate predictions, and it is very slow.

It turns out that just adding a cap on the beta norm fixed the problem. But
I'm keeping the cross validation code in here just in case.