Improved the way new models are added in EM
The old way to add a new model was to take a number of nearby noise
instances (where closeness was measured by Euclidean distance) and try to
fit a new model to it. Since Euclidean distance is often not correlated
with whether the instances come from the same model, this method often
generates many incorrect and unnecessary models.

The new method seeds a linear model with three adjacent instances, and then
tries to add every other noise instance to it. If adding an instance
increases the training set prediction error past the threshold
MODEL_ERROR_THRESH, then the instance is discarded. If the final model has
at least K instances, then it's considered a legitimate model and either
added as a new model or unified with an existing one. This approach is more
careful in finding instances generated by the same model. The rationale for
seeding the model with instances adjacent in time is that such instances
tend to come from the same model, due to the online nature of the training.