report MSVC target architecture
Report the MSVC target architecture during the build since Windows
users have to pick the correct one. The target architecture was already
being checked in order to set `g_msvs_variant`, so just move the logic
into a function, store it in a variable and use the result for both
actions. Helps with #165.