remove unused code related to lexer file stacks
`load_file` is the only function to call `stop_lex_from_file`, and it
does so right after the start function; `load_file` is never used, so
remove it and the stop function.

`start_lex_from_file`, then, is only called once, during lexer
initialization. `reading_from_top_level`, then, is always true, so
remove it and simplify all of the conditional blocks where it is used.