rename alt vars to lex vars
Rename the `alternate_input_...` variables to reflect the fact that
they are the only input for the lexer. It seems that previously files
were the main input to the lexer, and that the `alternate` variables
were a way to put off lexing the file and instead lex an input string.