remove input suffix
`lexer_input_suffix` was always provided so that the last lexeme given
by the lexer would be `R_PAREN_LEXEME`. The reason for this has probably
been gone for a long time (older Soar syntax?), but it was still required
because the parser and other code that relied on the lexer (the print
command and semantic memory) compensated by assuming that lexing would
finish with a right paren. Remove the input suffix and fix related code
so that lexing will always finish with EOF and client code will expect
it.