make determine_type_of_constituent_string private
This method only changes the type of `current_lexeme` and makes more
sense as a private method. The only place that used it was
`get_lexeme_from_string`; this function looked for quoted (`||`)
constants and then instantiated a `Lexer`. The lexer, however, will find
quoted symbolic constants from a simple call to `get_lexeme`, so this
code is extra and can be deleted in favor of `get_lexeme`.