Properly handle multiline errors in C++.
Added *EVENT_BAD = -1, to each of the enums that require an error handling.  This is the proper way to handle this instead of converting -1 to the 'enum' and then doing comparisons against -1 which the compiler will say will never exist.  This **might** mean that with optimizations this error code will never exist and will be optimized out.  This **might** lead to horrible behavior upon error codes.