modify spreading to directly add to access history
The problem was that every time history is used, it blew away the context activation.
Note that this is a problem unique to Soar - the spreading activation from WM buffers in ACT-R only affects retrieval bias, and does not contribute to the base-level of an item.

Additionally, the activation unification has been better integrated.
In particular, before, activating an element could lead to multiple accesses being recorded; this bug has been fixed (I think).