remove dead lex_alias code
Reroll of 5a8ce580ff.

Remove the code related to the `lex_alias` variable in agent.h,
including the `expansion_node` and `alias_struct` structs. This code
is uncommented (except for a date and originating file), and the logic
in `get_lexeme` is unreachable because `lex_alias` is always NIL. The
original functionality seems to have been to skip actual string lexing
and instead programatically provide a list of tokens, perhaps for
testing the parser.