remove top_dir_stack
Reroll of acd969576e.

The `top_dir_stack` variable in agent.h is not used anywhere, so
remove it along with related code and the `dir_stack_struct` struct.

I believe that this was not meant to be accessed by client code because
it's usage involves manual memory allocation and deallocation.