remove file usage in lexer
Reroll of 093b45996f.

Move `allow_ids` and `parentheses_level` into agent, and remove the
structs and logic for reading from files and stdin, which is no longer
done in the lexer. Also remove the memory deallocation required to
match the first call to `start_lex_from_file` inside of lexer.