store lexeme text in std::string
Reroll of ca021aa242.

There was a bug in the lexer due to it's use of a constant-length c
string. It checked the length of input quote and vertical bar lexemes
to make sure that they would fit within the range allocated for a
lexeme (`MAX_LEXEME_LENGTH`), but it didn't check other lexemes, so
it would accept, for example, a production name that was any length
and would simply write past the end of the array. On accident this is
unlikely to happen, but it is a risk.

Change `Lexeme.lex_string` to a `std::string` representation, and make
necessary changes in `Lexer` methods. For now, move `MAX_LINE_LENGTH`
into both `output_manager` and `sml_KernelSMLHandlers`, where it is
referenced. Note that the bug has only been moved from
`Lexeme` into other areas instead of being resolved (this will require
further attention).