allow compiling out SVS
Add a macro, `NO_SVS` which, if defined, prevents any reference to SVS from being used in the core code. Add a `--no-svs` option to Sconscript that defines the macro and prevents running the Sconscript in the SVS folder. Note that the macro is not yet documented, as I don't yet know a good place for that.