disambiguate floating points and dotted attributes
Disambiguate similar-looking floating points (`^num .1`) and
dotted attributes (`^num.1`) by checking the type of the last lexeme
and checking for a preceding space. Add documentation of the
disambiguation to `lex_period`.

Add a new private variable to the lexer, `prev_char`, to support the
disambiguation. Add/update/cleanup related documentation.

Closes #136.