fix all possible warnings
Fix all of the fixable warnings generated during a build:

`Command line warning D9025 : overriding '/W2' with '/w'`
was caused by trying to suppress warnings when building SWIG/TCL after
having set /W2 at the top level.

`decorated name length exceeded, name was truncated...` warnings were
caused by complicated structures in episodic_memory.h, and the warning
has to be suppressed at the end of the file since the warning is about
templates and they are expanded after the rest of the file is
processed.

Swig emitted `Warning 503: Can't wrap 'operator ()' unless renamed to
a valid identifier.` because of the `()` overload done in the
`cmp_str` struct in soar_instance.h, and was solved by instructing
Swig to rename it to `compare`.

Line 506 of explain.cpp was combaring a boolean with with a
`backtrace_str.result`, which is an int which if set to 1 indicates
that the structure instance is the result of a chunk. It might be
better to change the structure to hold a boolean instead of an int.

Warnings were emitted in semantic_memory.cpp because `temp_lti` and
`temp_const` were not initialized (in multiple places).

I did not touch the warning emitted by Export.h as it seems to be
purposeful; the warning is emitted when a non-static library is built.