Fix for bug that occurred when LTI changes level causing wmes removal
- When an LTI became disconnected and then was retrieved at a different
level than the one it was disconnected from, a buggy state could occur
that led to wme's with LTIs linked from that moved LTI to disappear.
Resetting the level to SMEM_LTI_UNKNOWN_VALUE when an LTI is
disconnected fixes the issue by forcing Soar to set the LTI's level as
based on what the caller to instantiate_rhs_value passes in.
- Added DT_UNKNOWN_LEVEL debug trace mode and thoroughly instrumented
unknown level code in decide.cpp