detect numeric attributes with periods
Previously the lexer would always read numbers as far as possible,
including trailing periods. This is awkward, though, when the period
is meant to specify a child attribute. Allow disambiguation via
whitespace: `1.1 a` is a float and a string, but `1.1a` is a number
with a string child `1a`. Do this by saving the lexer's place before
`read_rest_of_floating_point_number` and restoring it if no whitespace
is encountered before finding another constituent character.