quick fix for #245
When the CLI tokenizer was refactored between 9.2 and 9.3, the logic
for reading words enclosed in pipes was removed, though the
requirement that pipes match was kept inside of curly brace blocks.

Add back the capability to read pipe words, mostly using the existing
functionality for double-quotes. Pipes, unlike double quotes, need to
be saved in the output word in order to be used by the lexer.