Fix/hack for RL template issue
This is a hack to get around an RL template bug that surfaced
after we added identity-based STI variablization. For some
reason, the original template that created the instantiation
can now qualify as a duplicate of the instance, if that instance
had no conditions specialized by the match that created it.
Previously, bottom_node->first_child was null, indicating that
it was not a duplicate.  Spent several hours trying to figure out why,
but did not make progress.  Changed add_production_to_rete so that it
doesn't consider templates as a valid duplicate of a non-parsed rule
being added.  This seems to work for now, though it hasn't been
well-tested.  (very limited RL unit tests as of 3/2016).  Soar should
still detect duplicate templates that are added via the parser.