Fix for crash when trying to learn chunk with 2 unconnected goal states
- Problem occurred when original rule had two conditions with isa_goal
tests that didn't require the variables to be unequal
- If a match using the same goal is posible, it won't properly add a
isa_goal test to the second condition
- This caused that condition to not be identified as a root variable.
That made the condition and any below it seem unconnected, and the rule
repair code tries to fix it and crashes.
- Fixed it by changing how it marked which states it had already added
an isa_goal test to.  It now uses the variablized test rather than the
instantiated one, which isn't affected by the fact they matched the same
STI.