Repair improvements and start of trying to track down refcount issue
- Repair now generates identities and stores them in newly created
conditions
  - This fixes a crash that was occurring when another higher level
chunk was immediately built upon that rule that lacked identity in
repaired conditions.  Later matches would have correct identities, but
the ones that came from the recursive call to chunk_instantiation did
not.
  - To do this, repair now keeps a second map of symbols to identies.
- Action list now stores matched value to correspond to instantiated
conditions.  This is used to get the matched LTI that we need to
generate connecting conditions for.
- Path finding skips architectural WMEs now.  This prevents following
links up to other states and creating conditions based on other smem
retrievals in higher states.
- Improved failed chunk output
- Added some temp debug code to try to trace a particular symbol's
refcount