Fixed issue with re-orderer
It was splicing out new <@> test during initial parse of rule.  Even if
there was a positive conditions with the variable and then trying to
reverse it even.  To fix, I left those tests in the "simplified
conditions" so that they remain with the equality test.  It seems the
re-orderer once allowed multiple equality tests, so hopefully having
this test in the simplified conditions doesn't cause any issues.