Fix for long-standing GDS crash
I wrote a chunking agent that was able to crash in GDS code back to at
least 9.3.2.  It was trying to elaborate the GDS for the top level,
where a gds does not exist.  This occurs as a result of an identifier
created in the substate that is passed up to the top state as a result.
I'm not sure why that happens or what the proper solution is, but this
fix does prevent Soar from attempting to elaborate the top-state's gds
and then crashing.