Replaced backtracing algorithm with a much simpler version
It seems that most of the work that backtracing is doing to determine
which conditions are operational is not really necessary.  It seems to
me that we can determine whether a condition is accessible to a higher
level state simply by using the identifier level. This backtracing
algorithm eliminates the potential list, 2 of the tc numbers and the
numerous passes it performed to determine connectivity to higher states.

(Perhaps the identifier levels weren't being maintained when chunking
was originally written?  If not, I may be missing something.  This does
still pass all chunking unit tests, which are significant.)