Improved load_library and CLI clean-up/refactoring
- load library code will now search three paths for library, just like
settings and spawn debugger
  - Unfortunately, while this will allow it to find the tcl library in
another directory, it still can't load it.  The tcl library seems to
require that the soarlib is in the same directory, so this seems to be
something we could fix in tclsoarlib
- Some CLI clean-up
  - "cli" command now just "tcl" command
    - Took out a layer of the command, since it seems unlikely anyone
will be building similar interfaced.  So, instead of cli tcl on, it's
now simply tcl on/off
    - This actually involved refactoring some events that fire so that
they are less general.  More than you'd expect.
  - Combined cli file command into one cpp file.  Trying to get the cli
directory more manageable
  - Removed obsolete commands soarnews, edit-production
  - Move internal-symbols functionality into debug