Fixed some crashing caused by null RHS values that aren't zero
RHS values that are null can appear non-null because they have bit sets
to indicate which of the four types of rhs values.  Added a
rhs_value_is_null function that casts it to the appropriate type before
checking if the value is null.  Right now it's used in 2 or 3 places.
My guess is some other part of the code should not be recasting values
into specific rhs value types if it's null