Fixed tricky bug from prefs that retract b4 their insts are explained
Apparently, preferences_generated should have been called
preferences_generated_but_not_later_retracted.  This was causing
problems with the explainer and the visualizer because
preferences_generated was where it got some of the information needed
for the RHS.  So, not only was it missing information, sometimes it had
fewer preferences than actions, so things were mismatched.

The solution was tricky.  So, if the explainer is on and a preference in
a substate is retracted, Soar will now cache the preferences in the
instantiation until it retracts.  It removes the preference from the
appropriate lists and clears out some data so that it doesn't try to
clean up too much later.  This does mean increased memory use if the
explainer is on.

Also fixed a bug in the visualizer when it thought a rule in the
superstate was in the substate because the superstate happened to record
that rule for something else.  Visualizer now checks the match level of
the instantiation.