Fixed bug where opt connection used diff run phase interleave vs nonopt
Fixes issue #92.  Agent::RunSelf used a hard-coded sml_DECISION for
interleave parameter.  When in non-optimized mode, it would send it to
CLI which would use it for phse the stepsize and interleave parameters.
Optimized mode now just sends the stepsize for both parameters which
should be equivalent to what the CLI does.