Last epmem unit test should indeed fail in 9.6.0.  Disabled.
In Soar 9.6.0, this test should fail.  It was testing whether epmem
could determine that an LTI should only match after it was promoted to
an LTI. Epmem no longer tracks promotion time, so this should indeed
return an earlier episode.  Keeping around because it might make the
basis of a good test for epmem's new interaction with smem