Turned off duplicate name warning for now.  Source is a refcount bug.
Something is holding on to a refcount for a rule's name after an
init-soar. If Soar learns the exact same justifications the next run, it
will think that there's a name collision with the justifications from
the previous run, which causes it to generate a new name. This may
potentially cause a slow-down with agents that create many
justifications as well as a slow memory leak in the symbol pool.

Temporarily turning off that warning until we can track down the
refcount issue.