Fixed issue with rules being printed not being sourceable.
All symbols are now printed in a re-readable format. I'm not sure when
you'd not want them to not be re-readable, since it only seems to add
pipes when a symbol is ambiguous.  If it's an efficiency concern, the
cost may be much ameliorated by the fact that symbol strings are now
cached after the first time they are printed. (recent change in last
couple months)