Fixed big reorderer bug that caused instantiated chunk conds to be lost
We previously fixed the re-orderer so that it doesn't lose conditions,
but the original m_inst_top and m_inst_bottom weren't being updated
every time it tried to re-order or repair.  Now it does.

This bug was causing the instantiation to not be retracted, which
resulted in a memory leak.