Explainer/visualizer fixes
- Explainer will now correctly print identities for architecturally
created instantiations that don't have actual productions behind them.
They don't have original variables, but they do have variablization
identities as if they did.
- Fixed a visualizer crash that could occur when a condition tests a
child of a previous result.  The instantiation would not be part of the
backtrace, so it was crashing when trying to determine the shortest path
to a result.