Added more robust form of deep copy
Now adds conditions to the instantiation that test all of the original
items that were deep copied.  Any identifiers on the LHS will be
assigned an identity that will also be given to the new copied
identifier in the rhs preference.  So, deep copied items will now have
identities that are dependent on the reasoning that led to their
"parents".  (Constants are not given an identity.)