Added code to support EBC wrap linked STIs being created in a @ rhs func
- Added function wrap_with_lti_link()
- variablize_rhs changes
  - Added params pShouldLinkLTI, lti_link_tc
  - Calls wrap_with_lti_link()
- Updated RL template code to use it too when variablizing
- Added lti_link_function to ebc manager to cache the rhs function to 
  avoid lookup every time we wrap an LTI

Note:  While this did work, it broke identity unification of anything
backtracing through such wrapped functions.  While we could special case
this and consider the first argument of some RHS functions as the
"binding variable" and use it for unification, I think I'll just make it
a stand-alone RHS function in a new action.  Adding a new lti-linking
preference would also be essentially adding a new preference to the RHS,
so it seems that using a rhs function to link is still an easier option.