Removed unnecessary conjunctive test traversal by using eq_test
This is an attempt to improve performance.  Original soar code often
traverses conjunctive tests to find the equality test. During
development of EBC, we added an eq_test pointer that eliminated the need
to do this.  I did one final pass over the kernel looking for other
places that pointer could be used.  All of the functions that still did
it were very frequently used functions involved in instantiation
creation.  Many were in the RETE.
- bind_variables_in_test() in rete.cpp
- add_unbound_varnames_in_test() in rete.cpp
- var_test_bound_in_reconstructed_conds in rete.cpp
- var_bound_in_reconstructed_conds() in rete.cpp
- var_bound_in_reconstructed_conds() in rete.cpp
- cache_constraints_in_test() in ebc_constraints.cpp
- add_constraint_to_explanation() in explanation_trace.cpp
- test_contains_identity_in_set() in condition_record.cpp
- simplify_test() in production_reorder.cpp
- test_covered_by_bound_vars in production_reorder.cpp
- eliminated copy_of_equality_test_found_in_test() in test.cpp