Fixed a significant case that was not being literalized
A condition with an identity set that tested a literal rhs value was not
getting literalized.  The surprising thing is that I don't think it ever
has been.  EBC has only literalized the inverse case, i.e. a condition
with a literal test that matched a rhs value with an identity.  This fix
changes the results on several unit tests, but I think the new chunks
are the correct ones.  These are the tests affected:
	ChunkingTests::Constraint_Ungrounded
	ChunkingTests::Demo_Blocks_World_Hierarchical_Look_Ahead
	ChunkingTests::Demo_Blocks_World_Look_Ahead_State_Evaluation
	ChunkingTests::Demo_Blocks_World_Look_Ahead
	ChunkingTests::Demo_Eight_Puzzle
	ChunkingTests::Demo_MaC_Planning
	ChunkingTests::Demo_Water_Jug_Look_Ahead
	ChunkingTests::PRIMS_Sanity1
	ChunkingTests::PRIMS_Sanity2