Cache result in case we need to call twice
original commit messages:
* update to .i tcl
* rhs update

It is assumed that this function will only be called once, except in the
case that the return string buffer was too small and a re-call with a
larger one is needed. To prevent the actual function from being called
twice, we save the previous result to return immediately on the second
call. There is unfortunately no clear documentation on why this function
is sure to be called only once.