Fix Java version number parsing
The output of `java -version` appears to be formatted differently for
each distribution and differently from older versions. Therefore,
specify the distribution and version explicitly, and update the version
string parsing code so that it correctly gets the Java version instead
of the release date (the output from `java -version` is
`xyz version "11.0.10"`).

This fixes the error below:

NotImplementedError: Java version 2022-08-12 not supported:
  File "/home/runner/work/Soar/Soar/SConstruct", line 343:
  ...
  File "/home/runner/work/Soar/Soar/scons/scons-local-4.1.0/SCons/Tool/JavaCommon.py", line 105: