Add .editorconfig
The project is currently using a mix of tabs and spaces for indentation,
and the space indentation is not a uniform length. 4 spaces seems to be
the most common usage, so we'll specify that for all files except for
.soar files, which are currently using 3 spaces (and some tabs...).

Note that these settings do not affect the source code until applied by
an editor that supports .editorconfig. But they will help us to stay
consistent for future changes.