fix missing junit
Builds were failing because `junit` was not installed. Use a dedicated
action plugin to ensure it is available and also publish the test report
on PR's for convenient viewing. For now we don't create an actual PR
check, only annotations, because the check is currently sent to the
wrong build. See
https://github.com/mikepenz/action-junit-report/issues/40.