remove `using namespace std` everywhere
Attempt to fix this new build error:

error: reference to ‘byte’ is ambiguous

C++17 introduces `std::byte`, which is ambiguous with our own byte
definition when `using namespace std`. This particularly should never
be done in a header file. It was difficult to track down exactly which
usages were breaking the build, so I just removed them everywhere.