remove vfork
vfork is deprecated, as seen in the warning message:

    UnitTests/SoarUnitTests/FullTests.cpp:281:11: warning: 'vfork' is deprecated: Use posix_spawn or fork [-Wdeprecated-declarations]

I tried using plain `fork` instead, but this resulted in testHamilton
failing when sleep was not called. `spawn_posix` is the recommended
replacement for `vfork`, so I used that instead. The obvious syntax of
`char* argv = {"my", "string"}` triggers additional warnings because
those strings are constants but the declaration is not and `posix_spawn`
requires has a non-const in the signature. I declared plain char arrays
instead, as suggested here:
https://stackoverflow.com/questions/50596439/can-string-literals-be-passed-in-posix-spawns-argv

Another discussion for reference:
https://unix.stackexchange.com/questions/475609/why-is-vfork-intended-to-be-used-when-the-child-process-calls-exec-or-exit