Build and install Tcl from source on Windows
The install step is necessary because the build output structure is
non-deterministic. For example, on our last run, it put the lib file
under `.\win\Release_AMD64_VC1934\`, which will clearly change if the VC
version changes.

The build targets are not dependencies of the install target, so we need
to run `release` excplicitly.