Use proper atexit handler
Failing to shut down the Soar kernel before exit leads to a segfault,
so the safest thing to do is register an `atexit` handler that does
this. This script no longer raises any exceptions after the creation of
the kernel, but it will at least serve as an example of good practice.