svs_viewer builds on Mac
Follow the same pattern we used for Tcl, placing the environment setup
logic in a new file under build_support. Use the pkg-config file
included with GLFW. Several framework files are listed under
Libs.private, and I couldn't get pkg-config to emit these, so we still
have to manually specify them.

The viewer now builds on Mac, but still with plenty of warnings, since
OpenGL is entirely deprecated there and it's used heavily in the viewer
code.

Also, all I see is a black screen, so I'm not sure if it's
working. Running any of the example commands from the manual or from
tests.txt results in `unexpected field`.

Also, the code in the tests directory cannot be built because of a
missing 'foil.h'.

Also, the manual says to run this thing in the `out` directory, so maybe
it was only ever meant for people who built Soar themselves? Maybe I
shouldn't be trying to fix this build for distribution.

Related: #411.