Support --fired, --templates, --defaults
`production firing-counts --fired` didn't work because `--fired` was
not in the top-level options list for the production command.
`--template` and `--templates` were supported by different commands
with some inconsistent documentation; since we just find the longest
match for input, update to the plural form everywhere for consistency
(do the same for `--defaults`).

Add maintainer notes to the production command cli, as well as to the
one for WM, where I happened to notice several unnecessary options
declared at the top level.

Fixes #298.