Add `string` RHS function
Mostly this a copy of the trim function with most of the code deleted;
we don't need to do any manual copying or freeing of strings, since
`make_str_constant` does this for us.

Add a new category of unit tests, `BuiltinRHSTests`; I do see some
existing tests that exercise the built-in RHS functions, but they are
in the `ChunkingTests` category, and we're not testing chunks here.

Fixes #313.