Fix incorrect printa_sf calls everywhere
Because Soar's custom output formatter uses non-standard specifiers
(%u for uint64_t, %d for int64_t and %f for double), it's easy to get
the type input types incorrect, which leads to undefined behavior. For
example, in #342 Bob found that when printing the integer 0 with %d,
garbage would be output because an extra 4 bytes of memory were read
for the value. This could potentially lead to the program crashing
due to accessing restricted memory.

Fix these everywhere using `static_cast`. Also fix some places where
the specifiers were wrong in different ways (%u instead of %d, and %d
instead of %p in one case).

Fixes #342.