Preserve full decimal precision in string RHS
We add a new optional parameter to Symbol::to_string which specifies
the number of decimal places to use when converting a float to a string.
By default, this is 6, which is the same as used by fprintf, etc.
We only cache stringified floats when the default precision is used,
since that is the common case throughout the codebase.

Update the tests to check that a round-trip from float to string and
back to float again preserves the original value.