Use std::fixed to preserve precision
We want to preserve the old symbol float printing behavior, which always
had 6 digits after the decimal. To do this, we must set the formatting
to fixed. This also ensures that we are specifying the number of
digits *after* the decimal, not just the total digits.

Apply the same `std::fixed` fix to `AddDoubleInputWME`.

Add several more decimal points to the tests for the `string` RHS, now
that we can correctly round-trip them. 18 digits is more like it!