Update comments and tests around RHS buf re-alloc
Update incorrect and outdated comments on the handler defininitions
in sml_ClientEvents.h. Update the handlers in the tests to return
strings larger than the initially-allocated buffer so that the
re-allocation logic can be tested. Move the magic number to a #define
constant so it can be re-used in the test as well as for greater
clarity.

The comments about `extern "C"` no longer apply to the RHS and
ClientMessage handlers, since they now return `char*` again.