Don't redistribute Windows DLL's
We are building with Visual Studio 2022, and Microsoft has refactored/
reorganized/renamed the redistributable C++ runtime libs. See the docs:
https://learn.microsoft.com/en-us/cpp/windows/determining-which-dlls-to-redistribute
Our last release didn't include the DLL's simply because they weren't in
the location we expected to copy them from anymore. But the DLL's are
now only required when distributing for Windows before version 10; 10
and later come with the runtime libs pre-installed.

So: our last release would have broken Soar for any users that didn't
have the runtime installed on a distribution of Windows 8 or older.
No one has complained yet... though that isn't a guarantee. To support
them, we would have to re-write this logic to include the correct DLL's.
Since this code is currently broken, and is in the way of resolving
issue #322, I'll delete it for now. We can come back and write correct
DLL-inclusion code later if needed.