Specify script language
GH action workflows apparently issue `run` commands using `bash`, and
passing through `sudo` means that the script will be run using whatever
language the system thinks it is instead of simply `bash`. Without the
shebang line, all there is is the file extension, which causes the
script to be run as `sh` instead of `bash`, leading to this error:

./do_performance_test.sh: 23: [: full: unexpected operator
./do_performance_test.sh: 56: [: full: unexpected operator

Strangely, this only occurred on Linux, not on Mac.