Detect color support for SoarCLI
Double-clicking soar.exe on Windows opens it with CMD by default, which
doesn't support ANSI escape sequences. Detect whether colors are
supported in stdout, and use that to set the color option of SoarCLI.
This shouldn't change anything for POSIX users, but for Windows users
this will prevent printing garbage in CMD and will print colors in
Windows Terminal or Powershell. This also prevents printing ANSI escapes
when piping output to a file, though I'm not sure if anyone does that.