Remove bad -rpath argument from Tcl link command
Without a path argument, the first object file is added to the rpath and not
linked into the SO, leading to libtclsoarlib.so only containing symbols from
Tcl/src/main.cpp, and none from Tcl/src/TclSoarLib.cpp. I'm unsure of why this
setting was originally added and whether there used to be a real path argument.