Make extern "C" stuff in SML clearer
Move the `extern "C"` into the files where possible; that way we don't have an
extra `}` in the middle of the file that's disconnected syntactically from its
matching brace. I could not find the declaration for Tcl or Python, though
(and these use std::string directly still, so apparently no one ever cared about
making sure these had a C ABI in the DLL). We'll cross that bridge when we get
to it.

Add explanatory comments to the prevResult cacheing needed by
`RhsEventHandler`s and `ClientEventHandler`s.