Fix disappearing + in debugger in dark mode
Simply remove the background from the sidebar displayed behind the + buttons.
This was being explicitly set to white, while the + buttons change to white
in dark mode, so they were invisible. I tried using
`Display.isSystemDarkTheme()` to choose a different background, as well as
adding a border to it, but the minimal display really looks best. We allow
the canvas background to adjust to the color of the window, as the +'s do,
always giving good contrast and no visual noise.

Partial fix for #352.