harden smem export
Harden the logic for exporting smem to prevent segfaults. Previously we had a
segfault with just `smem -x file.soar` because the smem DB is not connected.
Adding an LTI option, however, would connect to the DB first. Change it so that
export (as with backup) requires connecting to the DB first, and gracefully
returns false if not connected.