connect to SMem DB in init CLI command
Previously a simple

smem --init
smem --clear

at the beginning of a source file would fail and prevent the rest of the file
from sourcing. Connect to the DB in the init CLI command so that any following
actions can access the DB. I believe this is the intuitive behavior, and I don't
see anything in the manual that would indicate that this incorrect.