Refactor to C++-style AddRhsHandler
Use `std::function` internally, converting to it in the C-style AddRhsHandler,
and add a new overload that accepts one directly. This is simpler to use for
C++ folks; it allows capturing pUserData in a function ahead of time, plays well
with with `std::bind` because there is no extra class instance pointer to keep
track of (previously the instance would have been passed in in pUserData).

Remove the extra pUserData storage everywhere. There are many more similar
refactorings that could be done, as well.