Fix warnings from CSharp SML bindings
Use `void*` or `uintptr_t` everywhere instead of `int`, which generates a
warning when converted to/from a pointer (because the size is wrong). Apply
macros for the different pointer contents everywhere.

Not merging for now because it needs to be tested, but I don't think I broke
anything. Whoever merges this should cherry-pick just this commit (and remove
this note, obviously).