Roll back 9.6.1 RhsHandler and ClientMessageHandler change
Ultimately it's not possible to simply use C-types in SML and get a portable
lib out of it, since C++ classes are used extensively. Creating a C-wrapper for
libSoar would be a large project of its own, and using the C++ code as-is is not
really viable at this point; we would need a SWIG-like tool (why on earth does
SWIG not support C++ to C wrapping?).

Fix and remove outstanding TODOs from the whole refactoring. Leave the new APIs
in place, as they are more ergonomic C++, anyway. Remove all logic related to
the buffer resizing, and remove duplicate code where possible.

Add documentation to the new methods.