improve invalid prod. diagnostics, untangle ebc
It seems like the reording logic may have lived in ebc at one time, though I
can't be sure about that. The logic that checks and warns for ungrounded
productions generated warnings related to chunking and used chunking trace
settings, even though it's fully possible to trigger these from regular `sp`
commands (and there is an additional call in RL, as well).

Move all of the chunking-related warnings and settings back to ebc. Instead of
setting `m_failure_type` on the agent's chunking manager, return the error code
from `reorder_and_validate_lhs_and_rhs` and check that in ebc to trigger logging
and agent stopping logic. Place the logic moved there into a new dedicated
method. Update other consumers of this method to check for `reorder_success`
instead of `true`, as well.

Put the reorder failure logging behind `OM_WARNINGS`, which is on by default and
is (correctly) not specific to chunking/ebc.

Rename `EBCFailureType` and `display_ebc_error` to reflect that fact that they
they aren't chunking-specific.

Update unit tests to check the warnings printed for ungrounded productions.