Fix failed command being interpreted as success
When a CLI command returns false, it is meant to indicate that the command
failed. However, `agent->GetLastCommandLineResult()` would still return true
if the command returned false, requiring the client to instead parse out the
return string to see if the command succeeded or not.

The cause of this was that the tokenizer had two separate bits of state that
needed to be set for error conditions, and one was missed for command failures.
`current.invalidate()` was called but `error` was not set. This would cause
`sourceError` to be empty in `CommandLineInterface::Source`, leading to
`m_LastError` being empty in `GetLastResultSML`, which would then not set the
error in the SML response. Whew!

Two things to prevent this from happening again in the future:

* encapsulate the 2 bits of required state change into the new function
`set_error` in tokenizer.h, and
* default to an `Unknown error` string in `CommandLineInterface::Source` when
the return value of the command was false but no error string was provided.

I was afraid that this was a large behavioral change in Soar, since Soar could
perhaps continue processing after a failed command before. However, this turned
out not to be the case. This simple Soar file would not source any productions,
since the middle command failed, either before or after this change:

```soar
svs --enable
step
svs --disable
sp {foo
(state <s> ^superstate nil)
-->}
```

Fixes #477.